# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/auth/employer/signin": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Get a JWT via given credentials.",
                "description": "Log in an employer",
                "operationId": "api.auth.employer.signin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "remember"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email"
                                    },
                                    "password": {
                                        "description": "password"
                                    },
                                    "remember": {
                                        "description": "remember me field"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "test@me.com",
                                    "password": "pass1234",
                                    "remember": false
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employer has been logged in successfully"
                    },
                    "400": {
                        "description": "Invalid Credentials"
                    },
                    "401": {
                        "description": "Invalid email or password"
                    },
                    "422": {
                        "description": "The given data was invalid"
                    }
                }
            }
        }
    }
}